finish Progress, Achievement, UI Checkpoint for Personal Page

+++ MyApplication/app/src/main/java/com/example/myapplication/model/Review.java
@@ -5,11 +5,11 @@ public class Review {
     private String subtitle;
     private String content;
     private String time;
-    private float rating;
+    private int rating;
     private int likeCount;
     private int[] imageResIds;
 
-    public Review(String name, String subtitle, String content, String time, float rating, int likeCount, int[] imageResIds) {
+    public Review(String name, String subtitle, String content, String time, int rating, int likeCount, int[] imageResIds) {
         this.name = name;
         this.subtitle = subtitle;
         this.content = content;
@@ -23,7 +23,7 @@ public class Review {
     public String getSubtitle() { return subtitle; }
     public String getContent() { return content; }
     public String getTime() { return time; }
-    public float getRating() { return rating; }
+    public int getRating() { return rating; }
     public int getLikeCount() { return likeCount; }
     public int[] getImageResIds() { return imageResIds; }
 }
